feat: toggle frontend to support multi difficulty
